rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-frontmatter==1.0.0
sentence-transformers==2.2.2
qdrant-client==1.7.0
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0